make price/amount u64

+++ src/store/mod.rs
@@ -1,3 +1,3 @@
 pub mod account;
-pub type Price = f64;
-pub type Amount = f64;
+pub type Price = u64;
+pub type Amount = u64;
